+++ database/migrations/2021_08_24_091551_create_order_details_table.php
@@ -15,8 +15,8 @@ class CreateOrderDetailsTable extends Migration
     {
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('item_id');
-            $table->integer('order_id');
+            $table->unsignedBigInteger('item_id');
+            $table->unsignedBigInteger('order_id');
             $table->unsignedDecimal('price');
             $table->timestamps();
         });
