+++ database/migrations/2021_08_29_092617_add_creditor_order_id_field_and_add_nullable_to_order_id_in_orders_details_table.php
@@ -14,8 +14,8 @@ class AddCreditorOrderIdFieldAndAddNullableToOrderIdInOrdersDetailsTable extends
     public function up()
     {
         Schema::table('order_details', function (Blueprint $table) {
-            $table->integer('creditor_order_id')->nullable();
-            $table->integer('order_id')->nullable()->change();
+            $table->unsignedBigInteger('creditor_order_id')->nullable();
+            $table->unsignedBigInteger('order_id')->nullable()->change();
         });
     }
