+++ database/migrations/2021_08_29_114634_add_company_id_company_name_and_notes_in_creditor_orders_table.php
@@ -14,7 +14,7 @@ class AddCompanyIdCompanyNameAndNotesInCreditorOrdersTable extends Migration
     public function up()
     {
         Schema::table('creditor_orders', function (Blueprint $table) {
-            $table->integer('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->string('company_name');
             $table->text('notes')->nullable();
         });
